refactor(api): register global hooks once and use emitter directly
