+++ api/lib/entities/memberships.service.js
@@ -1,6 +1,6 @@
 // @ts-check
 const { client: prisma } = require('../services/prisma.service');
-const hooks = require('../hooks');
+const hooks = require('../hooks/hookEmitter');
 
 /* eslint-disable max-len */
 /**
