+++ api/lib/entities/users.service.js
@@ -2,7 +2,7 @@
 const config = require('config');
 const elasticUsers = require('../services/elastic/users');
 const { client: prisma, Prisma } = require('../services/prisma.service');
-const hooks = require('../hooks');
+const hooks = require('../hooks/hookEmitter');
 
 const {
   MEMBER_ROLES: {
