+++ api/server.js
@@ -22,6 +22,12 @@ const ezreeportSync = require('./lib/services/sync/ezreeport');
 const elasticSync = require('./lib/services/sync/elastic');
 const kibanaSync = require('./lib/services/sync/kibana');
 
+/**
+ * Register hooks. Must not be called elsewhere. Some services can both
+ * trigger hooks and be used by hook handlers, leading to circular dependencies.
+ */
+require('./lib/hooks');
+
 const cronMetrics = require('./lib/controllers/metrics/cron');
 const { appLogger, httpLogger } = require('./lib/services/logger');
